backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.6.1
langchain==0.0.350
langchain-openai==0.0.2
chromadb==0.4.22
openpyxl==3.1.2
reportlab==4.0.8
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1